Function to dynamically validate page items added
This release extends ADC with an action type to dynamically validate any number of page items based on a centralized validation method.
It also changes the install script so that it does not throw away the existing installation but replaces only item which can be dynamically replaced.
It also introduces a new concept of action type owners to support multiple source of action types.